+++ src/main/java/com/lunark/lunark/auth/dto/AccountVerifiedDto.java
@@ -32,7 +32,6 @@ public class AccountVerifiedDto {
     public AccountVerifiedDto(Account account) {
         this.id = account.getId();
         this.email = account.getEmail();
-        this.password = account.getPassword();
         this.name = account.getName();
         this.surname = account.getSurname();
         this.address = account.getAddress();
