+++ src/main/java/com/lunark/lunark/auth/repository/ILdapAccountRepository.java
@@ -9,4 +9,5 @@ import java.util.UUID;
 
 public interface ILdapAccountRepository extends LdapRepository<LdapAccount> {
     Optional<LdapAccount> findByUuid(UUID uuid);
+    Optional<LdapAccount> findByEmail(String email);
 }
