+++ src/main/java/com/lunark/lunark/security/CustomUserDetailsService.java
@@ -1,5 +1,6 @@
 package com.lunark.lunark.security;
 
+import com.lunark.lunark.auth.repository.ILdapAccountRepository;
 import com.lunark.lunark.auth.service.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -9,12 +10,12 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 public class CustomUserDetailsService implements UserDetailsService {
 
     @Autowired
-    private AccountService accountService;
+    ILdapAccountRepository ldapAccountRepository;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         System.out.println(email);
-        return accountService.find(email)
+        return ldapAccountRepository.findByEmail(email)
                 .orElseThrow(() -> new UsernameNotFoundException("User not found with username " + email));
     }
 }
